ajout d'erreur au niveau du controller et correctif pour le LanguageManager et modification d'une station

+++ src/main/java/com/spring/crud/demo/controller/TrainStationController.java
@@ -11,8 +11,10 @@ import com.spring.crud.demo.model.TrainStation;
 import com.spring.crud.demo.repository.JourneyRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.spring.crud.demo.service.TrainStationService;
 
@@ -39,6 +41,11 @@ public class TrainStationController {
 	
 	@PutMapping("/{id}")
     public ResponseEntity<?> update(@PathVariable int id, @RequestBody TrainStation trainStation) {
+		String errorMessage = this.checkStation(id, "departure");
+		if (errorMessage != null) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, errorMessage);
+		}
+		
 		TrainStation updatedTrainStation = trainStationService.update(id, trainStation);
         return ResponseEntity.ok().body(updatedTrainStation);
     }
@@ -46,8 +53,43 @@ public class TrainStationController {
 	@DeleteMapping("/{id}")
 	@Transactional
     public ResponseEntity<?> delete(@PathVariable int id) {
+		String errorMessage = this.checkStation(id, "departure");
+		if (errorMessage != null) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, errorMessage);
+		}
+		
 		journeyRepository.deleteJourneysOfStation(id);
 		trainStationService.delete(id);
         return ResponseEntity.ok().body(LanguageManager.languageSelected.getString("station.deleted"));
     }
+	
+	public void checkBothStationError(int idFrom, int idTo) {
+		String errorMessage = "";
+
+		String error = checkStation(idFrom, "departure");
+		// checking if stations exist
+		if (error != null) {
+			errorMessage += error;
+		}
+		
+		error = checkStation(idTo, "arrival");
+		if (error != null) {
+			errorMessage += !errorMessage.equals("") ? "\n"+error : error;
+		}
+
+		// throwing error if one of the station don't exist
+		if (!errorMessage.equals("")) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, errorMessage);
+		}
+	}
+	
+	public String checkStation(int idStation, String typeStation) {
+		TrainStation station = trainStationService.checkIfStationExist(idStation);
+		String errorMessage = null;
+
+		if (station == null)
+			errorMessage = LanguageManager.languageSelected.getString("station." + typeStation + ".dontexist");
+
+		return errorMessage;
+	}
 }
